moved Debug components to debug folder
