+++ view/desktop/src/parts/TabbedPane/DebugComponents/DockviewControls.tsx
