+++ view/desktop/src/parts/TabbedPane/DebugComponents/LogsPanel.tsx
