+++ view/desktop/src/parts/TabbedPane/DebugComponents/Metadata.tsx
