+++ view/desktop/src/parts/TabbedPane/DebugComponents/debugPanel.tsx
