+++ view/desktop/src/parts/TabbedPane/DebugComponents/groupActions.tsx
