+++ view/desktop/src/parts/TabbedPane/DebugComponents/panelActions.tsx
